// types.ts
export interface SiteDefinition {
  id: string; // Ambiflo ID (CLLI)
  customerId: string;
  name: string;
  lat: number;
  lng: number;
  digitizedStatus?: 'Digitized' | 'Not Digitized';
}